Consolidate memory services to single MemoryService
- Replace ScopedMemoryService with single MemoryService using IScopedDatabaseService
- Update MCP tool tests to use consolidated service with proper DI setup
- Remove dual memory service pattern for cleaner architecture
- All memory-related tests still passing (10/10) with unified implementation

Eliminates confusion between multiple memory service implementations.
